% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleaningdata.R
\name{cleaning_data}
\alias{cleaning_data}
\title{Script to remove NA function}
\usage{
cleaning_data(filename)
}
\arguments{
\item{filename}{Name of the file to load in.}
}
\value{
Data frame with no NAs
}
\description{
Script to remove NA function
}
