% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jitterplot.R
\name{create_plot}
\alias{create_plot}
\title{Script to plot a jitter plot using ggplot2}
\usage{
create_plot(filename)
}
\arguments{
\item{filename}{name of the file to load in}
}
\value{
Jitterplot from dataframe
}
\description{
Script to plot a jitter plot using ggplot2
}
