% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linearmodel.R
\name{linear_model}
\alias{linear_model}
\title{Script to generate a linear model from dataframe}
\usage{
linear_model(filename)
}
\arguments{
\item{filename}{name of the file to load in}
}
\value{
Linear model from dataframe
}
\description{
Script to generate a linear model from dataframe
}
